postgres INTEGER not BIGINT

+++ sps.sql
@@ -7,6 +7,7 @@ CREATE SCHEMA sps;
 --  sps    | opinions    | table | jerkey
 --  sps    | screedlines | table | jerkey
 --  sps    | screeds     | table | jerkey
+-- https://www.postgresql.org/docs/current/datatype-numeric.html
 
 CREATE TABLE sps.screeds (
     pubkey      VARCHAR(64) NOT NULL PRIMARY KEY,
@@ -17,16 +18,16 @@ CREATE TABLE sps.screeds (
 
 -- SELECT * FROM sps.opinions;
 CREATE TABLE sps.opinions (
-    id BIGSERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     opinion TEXT NOT NULL,
-    screed_count BIGINT,
+    screed_count INTEGER,
     updated_at   TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE sps.screedlines (
     id SERIAL PRIMARY KEY,
     screed_key VARCHAR(64) NOT NULL,
-    opinion_id BIGINT NOT NULL,
+    opinion_id INTEGER NOT NULL,
     FOREIGN KEY (screed_key) REFERENCES sps.screeds(pubkey),
     FOREIGN KEY (opinion_id) REFERENCES sps.opinions(id)
 );
